streamlit/requirements.txt
streamlit 
pandas 
kafka-python 
faker
matplotlib